[UPD] - Inscripciones a cursos
Los usuario ahora pueden registrarse a cursos, esto se guardará en la base de datos y le aparecerá las inscripciones que tenga cada curso al administrador, de ahí podrá cambiar el estatus de estas.

+++ app/Http/Controllers/TecnicaController.php
@@ -41,12 +41,15 @@ class TecnicaController extends Controller
     }
 
     public function borrarTecnica($id) {
-        try {
-            $tecnica = Tecnica::findOrFail($id);
+        $tecnica = Tecnica::find($id);
+
+        if ($tecnica) {
+            $tecnica->productos()->detach();
             $tecnica->delete();
-            return response()->json(['success' => 'Técnica eliminada con éxito'], 204);
-        } catch (\Exception $e) {
-            return response()->json(['message' => 'Error del servidor', 'error' => $e->getMessage()], 500);
+    
+            return response()->json(['success' => 'Técnica eliminada correctamente.']);
+        } else {
+            return response()->json(['error' => 'Técnica no encontrada'], 404);
         }
     }
